refactor(jans-auth-server): extracted request object processing into separate method
https://github.com/JanssenProject/jans/issues/1208
