+++ jans-auth-server/server/src/main/java/io/jans/as/server/authorize/ws/rs/AuthzRequestService.java
@@ -1,11 +1,40 @@
 package io.jans.as.server.authorize.ws.rs;
 
+import com.google.common.collect.Lists;
+import com.google.common.collect.Sets;
+import io.jans.as.common.model.common.User;
+import io.jans.as.common.model.registration.Client;
 import io.jans.as.model.authorize.AuthorizeErrorResponseType;
+import io.jans.as.model.common.ResponseMode;
+import io.jans.as.model.config.WebKeysConfiguration;
 import io.jans.as.model.configuration.AppConfiguration;
+import io.jans.as.model.crypto.AbstractCryptoProvider;
+import io.jans.as.model.crypto.encryption.BlockEncryptionAlgorithm;
+import io.jans.as.model.crypto.encryption.KeyEncryptionAlgorithm;
+import io.jans.as.model.crypto.signature.AlgorithmFamily;
+import io.jans.as.model.crypto.signature.SignatureAlgorithm;
 import io.jans.as.model.error.ErrorResponseFactory;
+import io.jans.as.model.jwe.Jwe;
+import io.jans.as.model.jwk.Algorithm;
+import io.jans.as.model.jwk.JSONWebKeySet;
+import io.jans.as.model.jwk.Use;
+import io.jans.as.model.jwt.Jwt;
+import io.jans.as.model.jwt.JwtClaimName;
+import io.jans.as.model.jwt.JwtHeader;
+import io.jans.as.model.jwt.JwtHeaderName;
+import io.jans.as.model.token.JsonWebResponse;
+import io.jans.as.model.util.JwtUtil;
 import io.jans.as.model.util.Util;
 import io.jans.as.persistence.model.Par;
+import io.jans.as.server.model.authorize.Claim;
+import io.jans.as.server.model.authorize.IdTokenMember;
+import io.jans.as.server.model.authorize.JwtAuthorizationRequest;
+import io.jans.as.server.model.authorize.ScopeChecker;
 import io.jans.as.server.par.ws.rs.ParService;
+import io.jans.as.server.service.ClientService;
+import io.jans.as.server.service.RedirectUriResponse;
+import io.jans.as.server.service.RequestParameterService;
+import io.jans.as.server.service.ServerCryptoProvider;
 import jakarta.ejb.Stateless;
 import jakarta.inject.Inject;
 import jakarta.inject.Named;
@@ -13,10 +42,17 @@ import jakarta.ws.rs.WebApplicationException;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import org.apache.commons.lang.StringUtils;
+import org.jetbrains.annotations.Nullable;
+import org.json.JSONObject;
 import org.slf4j.Logger;
 
-import java.util.Map;
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
+import java.security.PrivateKey;
+import java.util.HashMap;
+import java.util.Set;
 
+import static io.jans.as.model.util.StringUtils.implode;
 import static org.apache.commons.lang3.BooleanUtils.isTrue;
 
 /**
@@ -36,10 +72,27 @@ public class AuthzRequestService {
     private ErrorResponseFactory errorResponseFactory;
 
     @Inject
+    private AuthorizeRestWebServiceValidator authorizeRestWebServiceValidator;
+
+    @Inject
     private ParService parService;
 
+    @Inject
+    private AbstractCryptoProvider cryptoProvider;
 
-    public boolean processPar(AuthzRequest authzRequest, Map<String, String> customParameters) {
+    @Inject
+    private ScopeChecker scopeChecker;
+
+    @Inject
+    private RequestParameterService requestParameterService;
+
+    @Inject
+    private WebKeysConfiguration webKeysConfiguration;
+
+    @Inject
+    private ClientService clientService;
+
+    public boolean processPar(AuthzRequest authzRequest) {
         boolean isPar = Util.isPar(authzRequest.getRequestUri());
         if (!isPar && isTrue(appConfiguration.getRequirePar())) {
             log.debug("Server configured for PAR only (via requirePar conf property). Failed to find PAR by request_uri (id): {}", authzRequest.getRequestUri());
@@ -85,9 +138,258 @@ public class AuthzRequestService {
             authzRequest.setOriginHeaders(par.getAttributes().getOriginHeaders());
         if (StringUtils.isNotBlank(par.getAttributes().getUiLocales()))
             authzRequest.setUiLocales(par.getAttributes().getUiLocales());
-        if (!par.getAttributes().getCustomParameters().isEmpty())
-            customParameters.putAll(par.getAttributes().getCustomParameters());
+        if (!par.getAttributes().getCustomParameters().isEmpty()) {
+            if (authzRequest.getCustomParameters() == null) {
+                authzRequest.setCustomParameters(new HashMap<>());
+            }
+            authzRequest.getCustomParameters().putAll(par.getAttributes().getCustomParameters());
+        }
 
         return isPar;
+    }
+
+    @SuppressWarnings("java:S3776")
+    public void processRequestObject(AuthzRequest authzRequest, Client client, RedirectUriResponse redirectUriResponse, Set<String> scopes, User user) {
+        JwtAuthorizationRequest jwtRequest = null;
+
+        if (StringUtils.isBlank(authzRequest.getRequest()) && StringUtils.isBlank(authzRequest.getRequestUri())) {
+            return;
+        }
+
+        try {
+            jwtRequest = JwtAuthorizationRequest.createJwtRequest(authzRequest.getRequest(), authzRequest.getRequestUri(), client, redirectUriResponse, cryptoProvider, appConfiguration);
+
+            if (jwtRequest == null) {
+                throw authorizeRestWebServiceValidator.createInvalidJwtRequestException(redirectUriResponse, "Failed to parse jwt.");
+            }
+            authzRequest.setJwtRequest(jwtRequest);
+
+            if (StringUtils.isNotBlank(jwtRequest.getState())) {
+                authzRequest.setState(jwtRequest.getState());
+                redirectUriResponse.setState(authzRequest.getState());
+            }
+            if (appConfiguration.isFapi() && StringUtils.isBlank(jwtRequest.getState())) {
+                authzRequest.setState(""); // #1250 - FAPI : discard state if in JWT we don't have state
+                redirectUriResponse.setState("");
+            }
+
+            if (jwtRequest.getRedirectUri() != null) {
+                redirectUriResponse.getRedirectUri().setBaseRedirectUri(jwtRequest.getRedirectUri());
+            }
+
+            SignatureAlgorithm signatureAlgorithm = SignatureAlgorithm.fromString(jwtRequest.getAlgorithm());
+            if (Boolean.TRUE.equals(appConfiguration.getForceSignedRequestObject()) && signatureAlgorithm == SignatureAlgorithm.NONE) {
+                throw authorizeRestWebServiceValidator.createInvalidJwtRequestException(redirectUriResponse, "A signed request object is required");
+            }
+
+            // JWT wins
+            if (!jwtRequest.getScopes().isEmpty()) {
+                if (!scopes.contains("openid")) { // spec: Even if a scope parameter is present in the Request Object value, a scope parameter MUST always be passed using the OAuth 2.0 request syntax containing the openid scope value
+                    throw new WebApplicationException(Response
+                            .status(Response.Status.BAD_REQUEST)
+                            .entity(errorResponseFactory.getErrorAsJson(AuthorizeErrorResponseType.INVALID_SCOPE, authzRequest.getState(), "scope parameter does not contain openid value which is required."))
+                            .build());
+                }
+                Set<String> checkedScopes = scopeChecker.checkScopesPolicy(client, Lists.newArrayList(jwtRequest.getScopes()));
+
+                scopes.clear();
+                scopes.addAll(checkedScopes);
+            }
+            if (jwtRequest.getRedirectUri() != null && !jwtRequest.getRedirectUri().equals(authzRequest.getRedirectUri())) {
+                throw authorizeRestWebServiceValidator.createInvalidJwtRequestException(redirectUriResponse, "The redirect_uri parameter is not the same in the JWT");
+            }
+            if (StringUtils.isNotBlank(jwtRequest.getNonce())) {
+                authzRequest.setNonce(jwtRequest.getNonce());
+            }
+            if (StringUtils.isNotBlank(jwtRequest.getCodeChallenge())) {
+                authzRequest.setCodeChallenge(jwtRequest.getCodeChallenge());
+            }
+            if (StringUtils.isNotBlank(jwtRequest.getCodeChallengeMethod())) {
+                authzRequest.setCodeChallengeMethod(jwtRequest.getCodeChallengeMethod());
+            }
+            if (jwtRequest.getDisplay() != null && StringUtils.isNotBlank(jwtRequest.getDisplay().getParamName())) {
+                authzRequest.setDisplay(jwtRequest.getDisplay().getParamName());
+            }
+            if (!jwtRequest.getPrompts().isEmpty()) {
+                authzRequest.setPrompt(implode(jwtRequest.getPrompts(), " "));
+            }
+            if (jwtRequest.getResponseMode() != null) {
+                authzRequest.setResponseMode(jwtRequest.getResponseMode().getValue());
+                redirectUriResponse.getRedirectUri().setResponseMode(jwtRequest.getResponseMode());
+            }
+
+            checkIdTokenMember(authzRequest, redirectUriResponse, user, jwtRequest);
+            requestParameterService.getCustomParameters(jwtRequest, authzRequest.getCustomParameters());
+        } catch (WebApplicationException e) {
+            JsonWebResponse jwr = parseRequestToJwr(authzRequest.getRequest());
+            handleJwr(authzRequest, client, redirectUriResponse, jwr);
+            throw e;
+        } catch (Exception e) {
+            log.error("Invalid JWT authorization request. Message : " + e.getMessage(), e);
+            throw authorizeRestWebServiceValidator.createInvalidJwtRequestException(redirectUriResponse, "Invalid JWT authorization request");
+        }
+
+
+        // JARM
+        Set<ResponseMode> jwtResponseModes = Sets.newHashSet(ResponseMode.QUERY_JWT, ResponseMode.FRAGMENT_JWT, ResponseMode.JWT, ResponseMode.FORM_POST_JWT);
+        if (jwtResponseModes.contains(authzRequest.getResponseModeEnum())) {
+            JsonWebResponse jwe = parseRequestToJwr(authzRequest.getRequest());
+            fillRedirectUriResponseforJARM(redirectUriResponse, jwe, client);
+        }
+
+        // Validate JWT request object after JARM check, because we want to return errors well formatted (JSON/JWT).
+        authorizeRestWebServiceValidator.validateJwtRequest(authzRequest.getClientId(), authzRequest.getState(), authzRequest.getHttpRequest(), authzRequest.getResponseTypeList(), redirectUriResponse, jwtRequest);
+    }
+
+    private void handleJwr(AuthzRequest authzRequest, Client client, RedirectUriResponse redirectUriResponse, JsonWebResponse jwr) {
+        if (jwr == null) {
+            return;
+        }
+
+        String checkForAlg = jwr.getClaims().getClaimAsString("alg"); // to handle Jans Issue#310
+        if ("none".equals(checkForAlg)) {
+            throw new WebApplicationException(Response.status(Response.Status.BAD_REQUEST)
+                    .entity(errorResponseFactory.getErrorAsJson(
+                            AuthorizeErrorResponseType.INVALID_REQUEST_OBJECT, "",
+                            "The None algorithm in nested JWT is not allowed for FAPI"))
+                    .type(MediaType.APPLICATION_JSON_TYPE).build());
+        }
+        ResponseMode responseMode = ResponseMode.getByValue(jwr.getClaims().getClaimAsString("response_mode"));
+        if (responseMode == ResponseMode.JWT) {
+            authzRequest.setResponseMode(responseMode.getValue());
+            redirectUriResponse.getRedirectUri().setResponseMode(ResponseMode.JWT);
+            fillRedirectUriResponseforJARM(redirectUriResponse, jwr, client);
+            if (appConfiguration.isFapi()) {
+                authorizeRestWebServiceValidator.throwInvalidJwtRequestExceptionAsJwtMode(
+                        redirectUriResponse, "Invalid JWT authorization request",
+                        jwr.getClaims().getClaimAsString("state"), authzRequest.getHttpRequest());
+            }
+        }
+    }
+
+    private void checkIdTokenMember(AuthzRequest authzRequest, RedirectUriResponse redirectUriResponse, User user, JwtAuthorizationRequest jwtRequest) {
+        final IdTokenMember idTokenMember = jwtRequest.getIdTokenMember();
+        if (idTokenMember == null) {
+            return;
+        }
+
+        if (idTokenMember.getMaxAge() != null) {
+            authzRequest.setMaxAge(idTokenMember.getMaxAge());
+        }
+        final Claim acrClaim = idTokenMember.getClaim(JwtClaimName.AUTHENTICATION_CONTEXT_CLASS_REFERENCE);
+        if (acrClaim != null && acrClaim.getClaimValue() != null) {
+            authzRequest.setAcrValues(acrClaim.getClaimValue().getValueAsString());
+        }
+
+        Claim userIdClaim = idTokenMember.getClaim(JwtClaimName.SUBJECT_IDENTIFIER);
+        if (userIdClaim != null && userIdClaim.getClaimValue() != null
+                && userIdClaim.getClaimValue().getValue() != null) {
+            String userIdClaimValue = userIdClaim.getClaimValue().getValue();
+
+            if (user != null) {
+                String userId = user.getUserId();
+
+                if (!userId.equalsIgnoreCase(userIdClaimValue)) {
+                    throw new WebApplicationException(redirectUriResponse.createErrorBuilder(AuthorizeErrorResponseType.USER_MISMATCHED).build());
+                }
+            }
+        }
+
+    }
+
+    @Nullable
+    public JsonWebResponse parseRequestToJwr(String request) {
+        if (request == null) {
+            return null;
+        }
+        String[] parts = request.split("\\.");
+        try {
+            if (parts.length == 5) {
+                String encodedHeader = parts[0];
+                JwtHeader jwtHeader = new JwtHeader(encodedHeader);
+                String keyId = jwtHeader.getKeyId();
+                PrivateKey privateKey = null;
+                KeyEncryptionAlgorithm keyEncryptionAlgorithm = KeyEncryptionAlgorithm
+                        .fromName(jwtHeader.getClaimAsString(JwtHeaderName.ALGORITHM));
+                if (AlgorithmFamily.RSA.equals(keyEncryptionAlgorithm.getFamily())) {
+                    privateKey = cryptoProvider.getPrivateKey(keyId);
+                }
+                return Jwe.parse(request, privateKey, null);
+            }
+            return Jwt.parseSilently(request);
+        } catch (Exception e) {
+            log.error(e.getMessage(), e);
+            return null;
+        }
+    }
+
+    private void fillRedirectUriResponseforJARM(RedirectUriResponse redirectUriResponse, JsonWebResponse jwr, Client client) {
+        try {
+            if (jwr != null) {
+                String tempRedirectUri = jwr.getClaims().getClaimAsString("redirect_uri");
+                if (StringUtils.isNotBlank(tempRedirectUri)) {
+                    redirectUriResponse.getRedirectUri().setBaseRedirectUri(URLDecoder.decode(tempRedirectUri, "UTF-8"));
+                }
+            }
+            String clientId = client.getClientId();
+            redirectUriResponse.getRedirectUri().setIssuer(appConfiguration.getIssuer());
+            redirectUriResponse.getRedirectUri().setAudience(clientId);
+            redirectUriResponse.getRedirectUri().setAuthorizationCodeLifetime(appConfiguration.getAuthorizationCodeLifetime());
+            redirectUriResponse.getRedirectUri().setSignatureAlgorithm(SignatureAlgorithm.fromString(client.getAttributes().getAuthorizationSignedResponseAlg()));
+            redirectUriResponse.getRedirectUri().setKeyEncryptionAlgorithm(KeyEncryptionAlgorithm.fromName(client.getAttributes().getAuthorizationEncryptedResponseAlg()));
+            redirectUriResponse.getRedirectUri().setBlockEncryptionAlgorithm(BlockEncryptionAlgorithm.fromName(client.getAttributes().getAuthorizationEncryptedResponseEnc()));
+            redirectUriResponse.getRedirectUri().setCryptoProvider(cryptoProvider);
+
+            String keyId = null;
+            if (client.getAttributes().getAuthorizationEncryptedResponseAlg() != null
+                    && client.getAttributes().getAuthorizationEncryptedResponseEnc() != null) {
+                if (client.getAttributes().getAuthorizationSignedResponseAlg() != null) { // Signed then Encrypted
+                    // response
+                    SignatureAlgorithm signatureAlgorithm = SignatureAlgorithm
+                            .fromString(client.getAttributes().getAuthorizationSignedResponseAlg());
+
+                    String nestedKeyId = new ServerCryptoProvider(cryptoProvider).getKeyId(webKeysConfiguration,
+                            Algorithm.fromString(signatureAlgorithm.getName()), Use.SIGNATURE);
+
+                    JSONObject jsonWebKeys = JwtUtil.getJSONWebKeys(client.getJwksUri());
+                    redirectUriResponse.getRedirectUri().setNestedJsonWebKeys(jsonWebKeys);
+
+                    String clientSecret = clientService.decryptSecret(client.getClientSecret());
+                    redirectUriResponse.getRedirectUri().setNestedSharedSecret(clientSecret);
+                    redirectUriResponse.getRedirectUri().setNestedKeyId(nestedKeyId);
+                }
+
+                // Encrypted response
+                JSONObject jsonWebKeys = JwtUtil.getJSONWebKeys(client.getJwksUri());
+                if (jsonWebKeys != null) {
+                    keyId = new ServerCryptoProvider(cryptoProvider).getKeyId(JSONWebKeySet.fromJSONObject(jsonWebKeys),
+                            Algorithm.fromString(client.getAttributes().getAuthorizationEncryptedResponseAlg()),
+                            Use.ENCRYPTION);
+                }
+                String sharedSecret = clientService.decryptSecret(client.getClientSecret());
+                byte[] sharedSymmetricKey = sharedSecret.getBytes(StandardCharsets.UTF_8);
+                redirectUriResponse.getRedirectUri().setSharedSymmetricKey(sharedSymmetricKey);
+                redirectUriResponse.getRedirectUri().setJsonWebKeys(jsonWebKeys);
+                redirectUriResponse.getRedirectUri().setKeyId(keyId);
+            } else { // Signed response
+                SignatureAlgorithm signatureAlgorithm = SignatureAlgorithm.RS256;
+                if (client.getAttributes().getAuthorizationSignedResponseAlg() != null) {
+                    signatureAlgorithm = SignatureAlgorithm
+                            .fromString(client.getAttributes().getAuthorizationSignedResponseAlg());
+                }
+
+                keyId = new ServerCryptoProvider(cryptoProvider).getKeyId(webKeysConfiguration,
+                        Algorithm.fromString(signatureAlgorithm.getName()), Use.SIGNATURE);
+
+                JSONObject jsonWebKeys = JwtUtil.getJSONWebKeys(client.getJwksUri());
+                redirectUriResponse.getRedirectUri().setJsonWebKeys(jsonWebKeys);
+
+                String clientSecret = clientService.decryptSecret(client.getClientSecret());
+                redirectUriResponse.getRedirectUri().setSharedSecret(clientSecret);
+                redirectUriResponse.getRedirectUri().setKeyId(keyId);
+            }
+        } catch (Exception e) {
+            log.error(e.getMessage(), e);
+        }
     }
 }
